Changed Client Location & renamed functions
